requirements.txt
requests
pytest
jsonschema == 4.19.1
allure-pytest==2.13.2
